+++ tests/test_stac.py
@@ -82,6 +82,21 @@ extensions = {
     "up42-system:metadata_version": "metadata-version",
 }
 
+MAPPING = [
+    ("title", "up42-user:title"),
+    ("tags", "up42-user:tags"),
+    ("product_id", "up42-product:product_id"),
+    ("collection_name", "up42-product:collection_name"),
+    ("modality", "up42-product:modality"),
+    ("order_id", "up42-order:order_id"),
+    ("asset_id", "up42-system:asset_id"),
+    ("account_id", "up42-system:account_id"),
+    ("workspace_id", "up42-system:workspace_id"),
+    ("job_id", "up42-system:job_id"),
+    ("source", "up42-system:source"),
+    ("metadata_version", "up42-system:metadata_version"),
+]
+
 
 class TestUp42ExtensionProvider:
     @pytest.mark.parametrize("entity_class", [pystac.Item, pystac.Collection])
@@ -111,16 +126,40 @@ class TestUp42ExtensionProvider:
             ),
         ],
     )
-    def test_should_provide_up42_extensions(self, entity):
-        assert entity.up42.title == extensions["up42-user:title"]  # type: ignore
-        assert entity.up42.tags == extensions["up42-user:tags"]  # type: ignore
-        assert entity.up42.product_id == extensions["up42-product:product_id"]  # type: ignore
-        assert entity.up42.collection_name == extensions["up42-product:collection_name"]  # type: ignore
-        assert entity.up42.modality == extensions["up42-product:modality"]  # type: ignore
-        assert entity.up42.order_id == extensions["up42-order:order_id"]  # type: ignore
-        assert entity.up42.asset_id == extensions["up42-system:asset_id"]  # type: ignore
-        assert entity.up42.account_id == extensions["up42-system:account_id"]  # type: ignore
-        assert entity.up42.workspace_id == extensions["up42-system:workspace_id"]  # type: ignore
-        assert entity.up42.job_id == extensions["up42-system:job_id"]  # type: ignore
-        assert entity.up42.source == extensions["up42-system:source"]  # type: ignore
-        assert entity.up42.metadata_version == extensions["up42-system:metadata_version"]  # type: ignore
+    @pytest.mark.parametrize("mapping", MAPPING)
+    def test_should_provide_up42_extensions(self, entity, mapping):
+        assert getattr(entity.up42, mapping[0]) == extensions[mapping[1]]  # type: ignore
+
+    @pytest.mark.parametrize(
+        "entity",
+        [
+            pystac.Item(
+                id=str(uuid.uuid4()),
+                collection=str(uuid.uuid4()),
+                geometry=None,
+                bbox=None,
+                datetime=dt.datetime.now(),
+                properties=extensions.copy(),
+            ),
+            pystac.Collection(
+                id=str(uuid.uuid4()),
+                description="",
+                extent=pystac.Extent(
+                    spatial=pystac.SpatialExtent(bboxes=[[1.0, 2.0, 3.0, 4.0]]),
+                    temporal=pystac.TemporalExtent(intervals=[[dt.datetime.now(), None]]),
+                ),
+                extra_fields=extensions.copy(),
+            ),
+        ],
+    )
+    @pytest.mark.parametrize("mapping", MAPPING)
+    def test_should_set_up42_extension(self, entity, mapping):
+        new_value = "new-value"
+        setattr(entity.up42, mapping[0], new_value)
+
+        if isinstance(entity, pystac.Item):
+            entity_dict = entity.properties
+        else:
+            entity_dict = entity.extra_fields
+
+        assert entity_dict[mapping[1]] == new_value